+++ src/editor.js
@@ -129,7 +129,12 @@ export class AbstractEditor {
       this.notify()
     }
 
-    const displayMode = this.dependenciesFulfilled ? 'block' : 'none'
+    let displayMode = this.dependenciesFulfilled ? 'block' : 'none'
+
+    if (this.options.hidden) {
+      displayMode = 'none'
+    }
+
     if (wrapper.tagName === 'TD') {
       Object.keys(wrapper.childNodes).forEach(child => (wrapper.childNodes[child].style.display = displayMode))
     } else wrapper.style.display = displayMode
@@ -293,10 +298,13 @@ export class AbstractEditor {
 
   onMove () {}
 
-  getButton (text, icon, title) {
+  getButton (text, icon, title, args = []) {
     const btnClass = `json-editor-btn-${icon}`
     if (!this.iconlib) icon = null
     else icon = this.iconlib.getIcon(icon)
+
+    text = this.translate(text, args)
+    title = this.translate(title, args)
 
     if (!icon && title) {
       text = title
@@ -308,9 +316,12 @@ export class AbstractEditor {
     return btn
   }
 
-  setButtonText (button, text, icon, title) {
+  setButtonText (button, text, icon, title, args = []) {
     if (!this.iconlib) icon = null
     else icon = this.iconlib.getIcon(icon)
+
+    text = this.translate(text, args)
+    title = this.translate(title, args)
 
     if (!icon && title) {
       text = title