+++ src/editors/button.js
@@ -38,7 +38,7 @@ export class ButtonEditor extends AbstractEditor {
     this.input.addEventListener('click', options.action, false)
 
     if (this.schema.readOnly || this.schema.readonly || this.schema.template) {
-      this.always_disabled = true
+      this.disable(true)
       this.input.setAttribute('readonly', 'true')
     }