+++ src/editors/radio.js
@@ -46,7 +46,7 @@ export class RadioEditor extends SelectEditor {
     }
 
     if (this.schema.readOnly || this.schema.readonly) {
-      this.always_disabled = true
+      this.disable(true)
       for (let j = 0; j < this.radioGroup.length; j++) {
         this.radioGroup[j].disabled = true
       }