+++ src/editors/string.js
@@ -127,7 +127,7 @@ export class StringEditor extends AbstractEditor {
     } else if (this.options.input_width) this.input.style.width = this.options.input_width
 
     if (this.schema.readOnly || this.schema.readonly || this.schema.template) {
-      this.always_disabled = true
+      this.disable(true)
       this.input.setAttribute('readonly', 'true')
     }