+++ src/editors/table.js
@@ -31,6 +31,7 @@ export class TableEditor extends ArrayEditor {
     this.item_default = itemSchema.default || null
     this.item_has_child_editors = itemSchema.properties || itemSchema.items
     this.width = 12
+    this.array_controls_top = this.options.array_controls_top || this.jsoneditor.options.array_controls_top
     super.preBuild()
   }
 
@@ -54,6 +55,10 @@ export class TableEditor extends ArrayEditor {
       this.header.textContent = this.getTitle()
       this.title = this.theme.getHeader(this.header, this.getPathDepth())
       this.container.appendChild(this.title)
+      if (this.options.infoText) {
+        this.infoButton = this.theme.getInfoButton(this.options.infoText)
+        this.container.appendChild(this.infoButton)
+      }
       this.title_controls = this.theme.getHeaderButtonHolder()
       this.title.appendChild(this.title_controls)
       if (this.schema.description) {
@@ -71,7 +76,11 @@ export class TableEditor extends ArrayEditor {
 
     this.panel.appendChild(this.table)
     this.controls = this.theme.getButtonHolder()
-    this.panel.appendChild(this.controls)
+    if (this.array_controls_top) {
+      this.title.appendChild(this.controls)
+    } else {
+      this.panel.appendChild(this.controls)
+    }
 
     if (this.item_has_child_editors) {
       const ce = tmp.getChildEditors()
@@ -311,7 +320,7 @@ export class TableEditor extends ArrayEditor {
   }
 
   _createDeleteButton (i, holder) {
-    const button = this.getButton('', 'delete', this.translate('button_delete_row_title_short'))
+    const button = this.getButton('', 'delete', 'button_delete_row_title_short')
     button.classList.add('delete', 'json-editor-btntype-delete')
     button.setAttribute('data-i', i)
     button.addEventListener('click', e => {
@@ -336,7 +345,7 @@ export class TableEditor extends ArrayEditor {
   }
 
   _createCopyButton (i, holder) {
-    const button = this.getButton('', 'copy', this.translate('button_copy_row_title_short'))
+    const button = this.getButton('', 'copy', 'button_copy_row_title_short')
     button.classList.add('copy', 'json-editor-btntype-copy')
     button.setAttribute('data-i', i)
     button.addEventListener('click', e => {
@@ -356,7 +365,7 @@ export class TableEditor extends ArrayEditor {
   }
 
   _createMoveUpButton (i, holder) {
-    const button = this.getButton('', 'moveup', this.translate('button_move_up_title'))
+    const button = this.getButton('', 'moveup', 'button_move_up_title')
     button.classList.add('moveup', 'json-editor-btntype-move')
     button.setAttribute('data-i', i)
     button.addEventListener('click', e => {
@@ -377,7 +386,7 @@ export class TableEditor extends ArrayEditor {
   }
 
   _createMoveDownButton (i, holder) {
-    const button = this.getButton('', 'movedown', this.translate('button_move_down_title'))
+    const button = this.getButton('', 'movedown', 'button_move_down_title')
     button.classList.add('movedown', 'json-editor-btntype-move')
     button.setAttribute('data-i', i)
     button.addEventListener('click', e => {
@@ -409,10 +418,10 @@ export class TableEditor extends ArrayEditor {
         this.setVisibility(this.panel, this.collapsed)
         if (this.collapsed) {
           this.collapsed = false
-          this.setButtonText(e.currentTarget, '', 'collapse', this.translate('button_collapse'))
+          this.setButtonText(e.currentTarget, '', 'collapse', 'button_collapse')
         } else {
           this.collapsed = true
-          this.setButtonText(e.currentTarget, '', 'expand', this.translate('button_expand'))
+          this.setButtonText(e.currentTarget, '', 'expand', 'button_expand')
         }
       })
 
@@ -436,13 +445,13 @@ export class TableEditor extends ArrayEditor {
   }
 
   _createToggleButton () {
-    const button = this.getButton('', 'collapse', this.translate('button_collapse'))
+    const button = this.getButton('', 'collapse', 'button_collapse')
     button.classList.add('json-editor-btntype-toggle')
     return button
   }
 
   _createAddRowButton () {
-    const button = this.getButton(this.getItemTitle(), 'add', this.translate('button_add_row_title', [this.getItemTitle()]))
+    const button = this.getButton(this.getItemTitle(), 'add', 'button_add_row_title', [this.getItemTitle()])
     button.classList.add('json-editor-btntype-add')
     button.addEventListener('click', (e) => {
       e.preventDefault()
@@ -459,7 +468,7 @@ export class TableEditor extends ArrayEditor {
   }
 
   _createDeleteLastRowButton () {
-    const button = this.getButton(this.translate('button_delete_last', [this.getItemTitle()]), 'subtract', this.translate('button_delete_last_title', [this.getItemTitle()]))
+    const button = this.getButton('button_delete_last', 'subtract', 'button_delete_last_title', [this.getItemTitle()])
     button.classList.add('json-editor-btntype-deletelast')
     button.addEventListener('click', (e) => {
       e.preventDefault()
@@ -480,7 +489,7 @@ export class TableEditor extends ArrayEditor {
   }
 
   _createRemoveAllRowsButton () {
-    const button = this.getButton(this.translate('button_delete_all'), 'delete', this.translate('button_delete_all_title'))
+    const button = this.getButton('button_delete_all', 'delete', 'button_delete_all_title')
     button.classList.add('json-editor-btntype-deleteall')
     button.addEventListener('click', (e) => {
       e.preventDefault()