+++ src/editors/upload.js
@@ -189,6 +189,7 @@ export class UploadEditor extends AbstractEditor {
         this.preview.appendChild(img)
       }
       img.onerror = error => {
+        // eslint-disable-next-line no-console
         console.error('upload error', error, error.currentTarget)
       }
       img.src = this.container.querySelector('a').href
@@ -216,7 +217,7 @@ export class UploadEditor extends AbstractEditor {
       file.formattedSize = `${parseFloat((file.size / (1024 ** i)).toFixed(2))} ${['Bytes', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'][i]}`
     } else file.formattedSize = '0 Bytes'
 
-    const uploadButton = this.getButton('Upload', 'upload', 'Upload')
+    const uploadButton = this.getButton('button_upload', 'upload', 'button_upload')
     uploadButton.addEventListener('click', (event) => {
       event.preventDefault()