+++ src/themes/bootstrap4.js
@@ -123,13 +123,41 @@ export class bootstrap4Theme extends AbstractTheme {
     minusBtn.textContent = '-'
     plusBtn.textContent = '+'
 
+    const initialize = (input, min) => {
+      if (min) {
+        input.value = Number(min)
+      } else {
+        input.value = Number(input.value)
+      }
+      input.setAttribute('initialized', '1')
+    }
+
+    const min = input.getAttribute('min')
+    const max = input.getAttribute('max')
+
     minusBtn.addEventListener('click', () => {
-      input.stepDown()
+      if (!input.getAttribute('initialized')) {
+        initialize(input, min)
+      } else if (min) {
+        if (Number(input.value) > Number(min)) {
+          input.stepDown()
+        }
+      } else {
+        input.stepDown()
+      }
       trigger(input, 'change')
     })
 
     plusBtn.addEventListener('click', () => {
-      input.stepUp()
+      if (!input.getAttribute('initialized')) {
+        initialize(input, min)
+      } else if (max) {
+        if (Number(input.value) < Number(max)) {
+          input.stepUp()
+        }
+      } else {
+        input.stepUp()
+      }
       trigger(input, 'change')
     })
 
@@ -218,6 +246,7 @@ export class bootstrap4Theme extends AbstractTheme {
       if (window.jQuery && window.jQuery().tooltip) {
         window.jQuery(button).tooltip()
       } else {
+        // eslint-disable-next-line no-console
         console.warn('Could not find popper jQuery plugin of Bootstrap.')
       }
     } else if (this.options.tooltip === 'css') {
@@ -447,7 +476,7 @@ export class bootstrap4Theme extends AbstractTheme {
   addInputError (input, text) {
     if (!input.controlgroup) return
 
-    input.classList.add('is-invalid')
+    input.controlgroup.classList.add('is-invalid')
 
     if (!input.errmsg) {
       input.errmsg = document.createElement('p')
@@ -463,7 +492,7 @@ export class bootstrap4Theme extends AbstractTheme {
   removeInputError (input) {
     if (!input.errmsg) return
     input.errmsg.style.display = 'none'
-    input.classList.remove('is-invalid')
+    input.controlgroup.classList.remove('is-invalid')
   }
 
   getTabHolder (propertyName) {