+++ tests/codeceptjs/editors/array_test.js
@@ -1,3 +1,5 @@
+/* global Feature Scenario */
+
 var assert = require('assert');
 
 Feature('array');
@@ -8,7 +10,7 @@ Scenario('should have correct initial value', async (I) => {
   assert.equal(await I.grabValueFrom('.debug'), '[]');
 });
 
-Scenario('should trigger array (table) editing triggers', async (I) => {
+Scenario('should trigger array (table) editing triggers @retry', async (I) => {
   I.amOnPage('table-move-events.html');
   I.seeElement('[data-schemapath="root.0"]');
   I.seeElement('[data-schemapath="root.1"]');
@@ -21,7 +23,7 @@ Scenario('should trigger array (table) editing triggers', async (I) => {
 
   I.amAcceptingPopups();
   I.click('//button[contains(@class, "json-editor-btn-moveup") and @data-i="1"]');
-  I.seeInPopup('moveRow');
+  I.retry({ retries: 5, minTimeout: 500 }).seeInPopup('moveRow');
   I.acceptPopup();
   I.click('.get-value');
   value = await I.grabValueFrom('.debug');
@@ -29,7 +31,7 @@ Scenario('should trigger array (table) editing triggers', async (I) => {
 
   I.amAcceptingPopups();
   I.click('//button[contains(@class, "json-editor-btn-movedown") and @data-i="1"]');
-  I.seeInPopup('moveRow');
+  I.retry({ retries: 5, minTimeout: 500 }).seeInPopup('moveRow');
   I.acceptPopup();
   I.click('.get-value');
   value = await I.grabValueFrom('.debug');
@@ -37,7 +39,7 @@ Scenario('should trigger array (table) editing triggers', async (I) => {
 
   I.amAcceptingPopups();
   I.click('//button[contains(@class, "json-editor-btn-copy") and @data-i="2"]');
-  I.seeInPopup('copyRow');
+  I.retry({ retries: 5, minTimeout: 500 }).seeInPopup('copyRow');
   I.acceptPopup();
   I.click('.get-value');
   value = await I.grabValueFrom('.debug');
@@ -45,7 +47,7 @@ Scenario('should trigger array (table) editing triggers', async (I) => {
 
   I.amAcceptingPopups();
   I.click('.json-editor-btntype-add');
-  I.seeInPopup('addRow');
+  I.retry({ retries: 5, minTimeout: 500 }).seeInPopup('addRow');
   I.acceptPopup();
   I.click('.get-value');
   value = await I.grabValueFrom('.debug');
@@ -58,10 +60,10 @@ Scenario('should trigger array (table) editing triggers', async (I) => {
   // form field. Similar to the '.debug' field.
   I.amAcceptingPopups();
   I.click('.json-editor-btntype-deletelast');
-  I.seeInPopup('Are you sure you want to remove this node?');
+  I.retry({ retries: 5, minTimeout: 500 }).seeInPopup('Are you sure you want to remove this node?');
   I.acceptPopup();
   I.amAcceptingPopups();
-  I.seeInPopup('deleteRow');
+  I.retry({ retries: 5, minTimeout: 500 }).seeInPopup('deleteRow');
   I.acceptPopup();
   I.click('.get-value');
   value = await I.grabValueFrom('.debug');
@@ -70,10 +72,10 @@ Scenario('should trigger array (table) editing triggers', async (I) => {
   // This test will fail when using Puppeteer due to the way Puppeteer handles popups.
   I.amAcceptingPopups();
   I.click('.json-editor-btntype-deleteall');
-  I.seeInPopup('Are you sure you want to remove this node?');
+  I.retry({ retries: 5, minTimeout: 500 }).seeInPopup('Are you sure you want to remove this node?');
   I.acceptPopup();
   I.amAcceptingPopups();
-  I.seeInPopup('deleteAllRows');
+  I.retry({ retries: 5, minTimeout: 500 }).seeInPopup('deleteAllRows');
   I.acceptPopup();
   I.click('.get-value');
   value = await I.grabValueFrom('.debug');
@@ -873,9 +875,9 @@ Scenario('should work well with nested array editors', async (I) => {
 Scenario('should work well with selectize multiselect editors', async (I) => {
   I.amOnPage('array-selectize.html');
   I.click('Add item');
+  await I.seeElement('[data-schemapath="root.0"]');
   I.click('Add item');
-  I.seeElement('[data-schemapath="root.0"]');
-  I.seeElement('[data-schemapath="root.1"]');
+  await I.seeElement('[data-schemapath="root.1"]');
   I.click('.get-value');
   value = await I.grabValueFrom('.debug');
   // ensure defaults